refactor: consolidate document endpoints into frappe/core/api/document.py
First consolidation of the public API surface plan: move the whitelisted
document CRUD endpoints from frappe/client.py, plus unlock_document from
frappe/model/document.py, into frappe/core/api/document.py. All moved
endpoints are marked @frappe.public(group="Documents") and got their
missing type annotations and Sphinx docstrings.

Backward compatibility: is_whitelisted() checks function identity and cmd
paths resolve via attribute import, so plain import aliases at the old
locations keep every existing dotted path working. frappe.client aliases
are permanent (those paths are baked into external integrations
everywhere); the frappe.model.document alias is resolved lazily via
module __getattr__ to avoid a circular import.

get_js (deprecated) and get_time_zone (moves with the user/timezone
consolidation later) stay in frappe/client.py.
